Introduce AbstractDependencyResolvingClasspath in prevision of test suite specific classpath
Story: gradle/langos#127
